- cleanup some socket issues and also a possible int overflow

+++ source/shared_lib/include/platform/common/platform_common.h
@@ -122,7 +122,7 @@ private:
 
 	Uint32 lastStartCount;
 	Uint32 lastTickCount;
-	int32  lastResult;
+	int64  lastResult;
 	int32 lastMultiplier;
 	bool lastStopped;
